Initial refactor of GortClient
