+++ src/gort/gort.py
@@ -22,36 +22,38 @@ from types import TracebackType
 from typing import (
     TYPE_CHECKING,
     Any,
-    Callable,
-    ClassVar,
-    Generic,
     Literal,
-    Sequence,
     Type,
     TypeVar,
 )
 
 from lvmopstools.retrier import Retrier
-from packaging.version import Version
 from rich import pretty, traceback
 from rich.logging import RichHandler
 from typing_extensions import Self
 
-from clu.client import AMQPClient, AMQPReply
+from clu.client import AMQPClient
 from sdsstools.logger import SDSSLogger, get_logger
 from sdsstools.time import get_sjd
 
 from gort import config
-from gort.core import RemoteActor
+from gort.devices.ag import AGSet
+from gort.devices.core import GortDevice, GortDeviceSet
+from gort.devices.enclosure import Enclosure
+from gort.devices.guider import GuiderSet
+from gort.devices.nps import NPSSet
+from gort.devices.spec import SpectrographSet
+from gort.devices.telemetry import TelemetrySet
+from gort.devices.telescope import TelescopeSet
 from gort.enums import Event
 from gort.exceptions import ErrorCode, GortError
 from gort.observer import GortObserver
 from gort.pubsub import notify_event
 from gort.recipes import recipes as recipe_to_class
+from gort.remote import RemoteActor
 from gort.tile import Tile
 from gort.tools import (
     get_temporary_file_path,
-    kubernetes_list_deployments,
     kubernetes_restart_deployment,
     overwatcher_is_running,
     run_in_executor,
@@ -75,7 +77,7 @@ except NameError:
     IPYTHON_DEFAULT_HOOKS = []
 
 
-__all__ = ["GortClient", "Gort", "GortDeviceSet", "GortDevice"]
+__all__ = ["GortClient", "Gort"]
 
 
 DevType = TypeVar("DevType", bound="GortDeviceSet | GortDevice")
@@ -118,14 +120,6 @@ class GortClient(AMQPClient):
         use_rich_output: bool = True,
         log_file_path: str | pathlib.Path | Literal[False] | None = None,
     ):
-        from gort.devices.ag import AGSet
-        from gort.devices.enclosure import Enclosure
-        from gort.devices.guider import GuiderSet
-        from gort.devices.nps import NPSSet
-        from gort.devices.spec import SpectrographSet
-        from gort.devices.telemetry import TelemetrySet as TelemSet
-        from gort.devices.telescope import TelescopeSet as TelSet
-
         self.client_uuid = str(uuid.uuid4()).split("-")[0]
 
         self._console: Console
@@ -147,20 +141,6 @@ class GortClient(AMQPClient):
             password=password,
             log=log,
         )
-
-        self.actors: dict[str, RemoteActor] = {}
-
-        self.config = deepcopy(config)
-
-        self.__device_sets = []
-
-        self.ags = self.add_device(AGSet, self.config["ags"]["devices"])
-        self.guiders = self.add_device(GuiderSet, self.config["guiders"]["devices"])
-        self.telescopes = self.add_device(TelSet, self.config["telescopes"]["devices"])
-        self.nps = self.add_device(NPSSet, self.config["nps"]["devices"])
-        self.specs = self.add_device(SpectrographSet, self.config["specs"]["devices"])
-        self.enclosure = self.add_device(Enclosure, name="enclosure", actor="lvmecp")
-        self.telemetry = self.add_device(TelemSet, self.config["telemetry"]["devices"])
 
     def _prepare_logger(
         self,
@@ -335,62 +315,13 @@ class GortClient(AMQPClient):
         # handler.
         self._setup_async_exception_hooks()
 
-        override_overwatcher = getattr(self, "_override_overwatcher", None)
-        override_envvar = os.environ.get("GORT_OVERRIDE_OVERWATCHER", "0")
-        if override_overwatcher is None:
-            override_overwatcher = False if override_envvar == "0" else True
-
-        if override_overwatcher is None:
-            pass
-        else:
-            overwatcher_running = await overwatcher_is_running(self)
-            if overwatcher_running:
-                if override_overwatcher:
-                    self.log.warning("Overwatcher is running, be careful!")
-                else:
-                    raise GortError(
-                        "Overwatcher is running. If you really want to use GORT "
-                        "initialise it with Gort(override_overwatcher=True).",
-                        error_code=ErrorCode.OVERATCHER_RUNNING,
-                    )
-
-        await asyncio.gather(*[ractor.init() for ractor in self.actors.values()])
-
-        # Initialise device sets.
-        await asyncio.gather(*[dev.init() for dev in self.__device_sets])
-
         return self
-
-    def add_device(self, class_: Type[DevType], *args, **kwargs) -> DevType:
-        """Adds a new device or device set to Gort."""
-
-        ds = class_(self, *args, **kwargs)
-        self.__device_sets.append(ds)
-
-        return ds
 
     @property
     def connected(self):
         """Returns :obj:`True` if the client is connected."""
 
         return self.connection and self.connection.connection is not None
-
-    def add_actor(self, actor: str, device: GortDevice | None = None):
-        """Adds an actor to the programmatic API.
-
-        Parameters
-        ----------
-        actor
-            The name of the actor to add.
-        device
-            A device associated with this actor.
-
-        """
-
-        if actor not in self.actors:
-            self.actors[actor] = RemoteActor(self, actor, device=device)
-
-        return self.actors[actor]
 
     def set_verbosity(self, verbosity: str | int | None = None):
         """Sets the level of verbosity to ``debug``, ``info``, or ``warning``.
@@ -413,283 +344,9 @@ class GortClient(AMQPClient):
         if verbosity not in ["debug", "info", "warning"]:
             raise ValueError("Invalid verbosity value.")
 
-        verbosity_level = logging.getLevelName(verbosity.upper())
-        self.log.sh.setLevel(verbosity_level)
-
-
-GortDeviceType = TypeVar("GortDeviceType", bound="GortDevice")
-
-
-class GortDeviceSet(dict[str, GortDeviceType], Generic[GortDeviceType]):
-    """A set to gort-managed devices.
-
-    Devices can be accessed as items of the :obj:`.GortDeviceSet` dictionary
-    or using dot notation, as attributes.
-
-    Parameters
-    ----------
-    gort
-        The :obj:`.GortClient` instance.
-    data
-        A mapping of device to device info. Each device must at least include
-        an ``actor`` key with the actor to use to communicated with the device.
-        Any other information is passed to the :obj:`.GortDevice` on instantiation.
-    kwargs
-        Other keyword arguments to pass wo the device class.
-
-    """
-
-    __DEVICE_CLASS__: ClassVar[Type["GortDevice"]]
-    __DEPLOYMENTS__: ClassVar[list[str]] = []
-
-    def __init__(self, gort: GortClient, data: dict[str, dict], **kwargs):
-        self.gort = gort
-
-        _dict_data = {}
-        for device_name in data:
-            device_data = data[device_name].copy()
-            actor_name = device_data.pop("actor")
-            _dict_data[device_name] = self.__DEVICE_CLASS__(
-                gort,
-                device_name,
-                actor_name,
-                **device_data,
-                **kwargs,
-            )
-
-        dict.__init__(self, _dict_data)
-
-    async def init(self):
-        """Runs asynchronous tasks that must be executed on init."""
-
-        # Run devices init methods.
-        results = await asyncio.gather(
-            *[dev.init() for dev in self.values()],
-            return_exceptions=True,
-        )
-
-        for idev, result in enumerate(results):
-            if isinstance(result, Exception):
-                self.write_to_log(
-                    f"Failed initialising device {list(self)[idev]} "
-                    f"with error {str(result)}",
-                    "error",
-                )
-
-        return
-
-    def __getattribute__(self, __name: str) -> Any:
-        if __name in self:
-            return self.__getitem__(__name)
-        return super().__getattribute__(__name)
-
-    async def call_device_method(self, method: Callable, *args, **kwargs):
-        """Calls a method in each one of the devices.
-
-        Parameters
-        ----------
-        method
-            The method to call. This must be the abstract class method,
-            not the method from an instantiated object.
-        args,kwargs
-            Arguments to pass to the method.
-
-        """
-
-        if not callable(method):
-            raise GortError("Method is not callable.")
-
-        if hasattr(method, "__self__"):
-            # This is a bound method, so let's get the class method.
-            method = method.__func__
-
-        if not hasattr(self.__DEVICE_CLASS__, method.__name__):
-            raise GortError("Method does not belong to this class devices.")
-
-        devices = self.values()
-
-        return await asyncio.gather(*[method(dev, *args, **kwargs) for dev in devices])
-
-    async def send_command_all(
-        self,
-        command: str,
-        *args,
-        devices: Sequence[str] | None = None,
-        **kwargs,
-    ):
-        """Sends a command to all the devices.
-
-        Parameters
-        ----------
-        command
-            The command to call.
-        args, kwargs
-            Arguments to pass to the :obj:`.RemoteCommand`.
-
-        """
-
-        tasks = []
-        for name, dev in self.items():
-            if devices is not None and name not in devices:
-                continue
-
-            actor_command = dev.actor.commands[command]
-            tasks.append(actor_command(*args, **kwargs))
-
-        return await asyncio.gather(*tasks)
-
-    def write_to_log(
-        self,
-        message: str,
-        level: str = "debug",
-        header: str | None = None,
-    ):
-        """Writes a message to the log with a custom header.
-
-        Parameters
-        ----------
-        message
-            The message to log.
-        level
-            The level to use for logging: ``'debug'``, ``'info'``, ``'warning'``, or
-            ``'error'``.
-        header
-            The header to prepend to the message. By default uses the class name.
-
-        """
-
-        if header is None:
-            header = f"({self.__class__.__name__}) "
-
-        message = f"{header}{message}"
-
-        level = logging.getLevelName(level.upper())
-        assert isinstance(level, int)
-
-        self.gort.log.log(level, message)
-
-    async def restart(self):
-        """Restarts the set deployments and resets all controllers.
-
-        Returns
-        -------
-        result
-            A boolean indicting if the entire restart procedure succeeded.
-
-        """
-
-        failed: bool = False
-
-        self.write_to_log("Restarting Kubernetes deployments.", "info")
-        for deployment in self.__DEPLOYMENTS__:
-            await kubernetes_restart_deployment(deployment)
-
-        self.write_to_log("Waiting 15 seconds for deployments to be ready.", "info")
-        await asyncio.sleep(15)
-
-        # Check that deployments are running.
-        running_deployments = await kubernetes_list_deployments()
-        for deployment in self.__DEPLOYMENTS__:
-            if deployment not in running_deployments:
-                failed = True
-                self.write_to_log(f"Deployment {deployment} did not restart.", "error")
-
-        # Refresh the command models for all the actors.
-        await asyncio.gather(*[actor.refresh() for actor in self.gort.actors.values()])
-
-        # Refresh the device set.
-        await self.init()
-
-        return not failed
-
-
-class GortDevice:
-    """A gort-managed device.
-
-    Parameters
-    ----------
-    gort
-        The :obj:`.GortClient` instance.
-    name
-        The name of the device.
-    actor
-        The name of the actor used to interface with this device. The actor is
-        added to the list of :obj:`.RemoteActor` in the :obj:`.GortClient`.
-
-
-    """
-
-    def __init__(self, gort: GortClient, name: str, actor: str):
-        self.gort = gort
-        self.name = name
-        self.actor = gort.add_actor(actor, device=self)
-
-        # Placeholder version. The real one is retrieved on init.
-        self.version = Version("0.99.0")
-
-    async def init(self):
-        """Runs asynchronous tasks that must be executed on init.
-
-        If the device is part of a :obj:`.DeviceSet`, this method is called
-        by :obj:`.DeviceSet.init`.
-
-        """
-
-        # Get the version of the actor.
-        if "version" in self.actor.commands:
-            try:
-                reply = await self.actor.commands.version()
-                if (version := reply.get("version")) is not None:
-                    self.version = Version(version)
-            except Exception:
-                pass
-
-        return
-
-    def write_to_log(
-        self,
-        message: str,
-        level: str = "debug",
-        header: str | None = None,
-    ):
-        """Writes a message to the log with a custom header.
-
-        Parameters
-        ----------
-        message
-            The message to log.
-        level
-            The level to use for logging: ``'debug'``, ``'info'``, ``'warning'``, or
-            ``'error'``.
-        header
-            The header to prepend to the message. By default uses the device name.
-
-        """
-
-        if header is None:
-            header = f"({self.name}) "
-
-        message = f"{header}{message}"
-
-        level = logging.getLevelName(level.upper())
-        assert isinstance(level, int)
-
-        self.gort.log.log(level, message)
-
-    def log_replies(self, reply: AMQPReply, skip_debug: bool = True):
-        """Outputs command replies."""
-
-        if reply.body:
-            if reply.message_code in ["w"]:
-                level = "warning"
-            elif reply.message_code in ["e", "f", "!"]:
-                level = "error"
-            else:
-                level = "debug"
-                if skip_debug:
-                    return
-
-            self.write_to_log(str(reply.body), level)
+        level_mapping = logging.getLevelNamesMapping()
+        if verbosity_level := level_mapping.get(verbosity.upper()):
+            self.log.sh.setLevel(verbosity_level)
 
 
 class Gort(GortClient):
@@ -725,6 +382,20 @@ class Gort(GortClient):
 
         super().__init__(*args, **kwargs)
 
+        self.actors: dict[str, RemoteActor] = {}
+
+        self.config = deepcopy(config)
+
+        self.__device_sets = []
+
+        self.ags = self.add_device(AGSet, config["ags.devices"])
+        self.guiders = self.add_device(GuiderSet, config["guiders.devices"])
+        self.telescopes = self.add_device(TelescopeSet, config["telescopes.devices"])
+        self.nps = self.add_device(NPSSet, config["nps.devices"])
+        self.specs = self.add_device(SpectrographSet, config["specs.devices"])
+        self.enclosure = self.add_device(Enclosure, name="enclosure", actor="lvmecp")
+        self.telemetry = self.add_device(TelemetrySet, config["telemetry.devices"])
+
         self._override_overwatcher = override_overwatcher
 
         self.observer = GortObserver(self)
@@ -732,6 +403,63 @@ class Gort(GortClient):
 
         if verbosity:
             self.set_verbosity(verbosity)
+
+    async def init(self) -> Self:
+        """Initialises the client and all devices."""
+
+        await super().init()
+
+        override_overwatcher = self._override_overwatcher
+        override_envvar = os.environ.get("GORT_OVERRIDE_OVERWATCHER", "0")
+
+        if override_overwatcher is None:
+            override_overwatcher = False if override_envvar == "0" else True
+
+        if override_overwatcher is None:
+            pass
+        else:
+            overwatcher_running = await overwatcher_is_running(self)
+            if overwatcher_running:
+                if override_overwatcher:
+                    self.log.warning("Overwatcher is running, be careful!")
+                else:
+                    raise GortError(
+                        "Overwatcher is running. If you really want to use GORT "
+                        "initialise it with Gort(override_overwatcher=True).",
+                        error_code=ErrorCode.OVERATCHER_RUNNING,
+                    )
+
+        await asyncio.gather(*[ractor.init() for ractor in self.actors.values()])
+
+        # Initialise device sets.
+        await asyncio.gather(*[dev.init() for dev in self.__device_sets])
+
+        return self
+
+    def add_device(self, class_: Type[DevType], *args, **kwargs) -> DevType:
+        """Adds a new device or device set to Gort."""
+
+        ds = class_(self, *args, **kwargs)
+        self.__device_sets.append(ds)
+
+        return ds
+
+    def add_actor(self, actor: str, device: GortDevice | None = None):
+        """Adds an actor to the programmatic API.
+
+        Parameters
+        ----------
+        actor
+            The name of the actor to add.
+        device
+            A device associated with this actor.
+
+        """
+
+        if actor not in self.actors:
+            self.actors[actor] = RemoteActor(self, actor, device=device)
+
+        return self.actors[actor]
 
     @Retrier(max_attempts=3, delay=3)
     async def emergency_shutdown(self):
